backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
lightrag-hku>=1.3.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
numpy>=1.26.0
networkx>=3.2.0
python-multipart>=0.0.20
nano-vectordb>=0.0.4
